+++ src/giterm/panels/__init__.py
@@ -1,0 +1,5 @@
+from .panel import Panel, Popup, StateLinePanel
+from .branches import BranchesPanel
+from .diff import DiffPanel
+from .history import HistoryPanel
+from .stager import StagerUnstagerPanel
